[controller] support new get election method
